+++ backend/app/ai-safety/src/utils/constants.py
@@ -1,0 +1,1 @@
+SLUR_LIST_FILENAME = "curated_slurlist_hi_en.csv"
