+++ backend/app/ai-safety/src/utils/language_detector.py
